requirements.txt
numpy==1.14.0
opencv-python==4.7.0.68
Pillow==4.1.1
